Fixed starting SEB by opening a seb:// link with custom .seb config didn't work. Now removing all persisted open webpages when quitting properly SEB. Fixed various build errors.

+++ Classes/Global/MyGlobals.m
@@ -45,7 +45,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(MyGlobals);
 + (int)ddLogLevel
 {
     ddLogLevel = [[self sharedMyGlobals] logLevel];
-    return ddLogLevel;
+    return (int)ddLogLevel;
 }
